requirements.txt
requests_unixsocket>=0.2.0
requests>=2.25.1
urllib3>=1.25.10 # for requests
tabulate==0.8.7
ffmpeg_python==0.2.0
siaskynet>=2.2.0
youtube_dl>=2021.6.6
opencv-python
base58>=2.0.1
secp256k1
cryptography
ipfshttpclient==0.8.0a2
pycryptodome>=3.14.1
beem>=0.24.20